10242: Add option set warnings
