+++ app/models/ability.rb
@@ -239,7 +239,7 @@ class Ability
 
     # we need these specialized permissions because option names/hints are updated via option set
     cannot(%i[add_options remove_options reorder_options], OptionSet, &:published?)
-    cannot(:destroy, OptionSet) { |o| o.data? || o.any_questions? || o.published? }
+    cannot(:destroy, OptionSet) { |o| o.data? || o.in_use? || o.published? }
 
     # operations can't be destroyed while their underlying job is in progress
     cannot :destroy, Operation do |op|
