+++ app/models/forms/integrity_warnings/question_warner.rb
@@ -16,6 +16,16 @@ module Forms
         # for standard copies.
         %i[published data standard_copy]
       end
+
+      def form_list
+        form_count = object.forms.size
+        more_suffix =
+          if form_count > MAX_FORMS
+            str = I18n.t("integrity_warnings.more_suffix", count: form_count - 3)
+            "(+#{str})"
+          end
+        [object.forms.by_name[0...MAX_FORMS].map(&:name).join(", "), more_suffix].compact.join(" ")
+      end
     end
   end
 end
